Remove name(space) setting and connection halting from CustomPlug
There were two problems with the plug:

The `Appsignal.Plug.set_name` and `Appsignal.Plug.set_namespace`
functions were called on a pid, not on an Appsignal.Span, so they were
ignored. This didn't actually cause the issue, but it didn't work as
intended.

The other issue wwas the halting of the response. By doing that, the
rest of the plugs aren't called anymore, so the parent span (the one
started by Phoenix in this example) is never closed. The child span,
which _is_ closed, therefor gets dropped.