Wait for child process to gracefully exit
When docker-compose is started as a child process, it continues to run
even after the rake task has exited. This is because it "traps" the
interrupt (`CTRL + C`) and Ruby doesn't wait for the process to really end.

Switch `system` usage to `spawn` like we also do for the agent
integration test suite. Connect it to the parent process STDIN, STDOUT
and STDERR so that it has the same behavior as `system`.

Except, wait for the process to really end. Use `Process.wait2` to wait
for the child process to quit. But also register the PID on the
`child_processes` array. Then loop over this array of process in our own
"trap" on interrupt and wait for the process to really end.

If a user presses `CTRL + C` again (another interrupt) docker-compose
will continue to "kill" the process instead and exit more quickly. The
rake task interrupt continues to wait until the process has truly exited
and only then exit itself.