Make default app port 4001 (#13)
Port 4001 should not conflict with most apps that can also run on the
host machine at the same time.

I also removed all the other `ports:` statements for databases from the
`docker-compose.yml` files. The `ports:` statements expose ports from
the container to the host server, and for databases this is not required
to function. All containers are on their own network and can communicate
with each other without explicitly exposing ports between containers.

Closes #10