Use docker compose to build images
Instead using using `docker build` and tagging an image manually, use
`docker-compose build` and let it tag the image based on the image name
defined in the `docker-compose.yml` file.

When I was adding a react app using the JavaScript integration I moved
it from the `nodejs` to the `javascript` directory, and didn't realize
the docker image was automatically tagged based on the directory name
and the image name referred to in the `docker-compose.yml` file didn't
match anymore. I was building a new image but still testing against the
old image.

This change moves all that config to the `docker-compose.yml` file and
doesn't base the image name on the app directory anymore, only during
`rake app:new` when the `docker-compose.yml` is generated. By specifying
`build: .` in the `docker-compose.yml` file we specify the context for
where the `Dockerfile` can be found.

I also removed the tag from the build process and image name. If none is
specified docker compose automatically uses the "latest" tag.