Add JavaScript webpack test app
Add basic JavaScript app that uses webpack to precompile the JavaScript
app.

Use the JavaScript integration along with the webpack plugin to send
private sourcemaps for the app.

It's a little non standard setup, but I cannot wait for the container to
restart when running the server in `boot.sh`. It takes 10+ minutes on my
machine to build all the integrations, link them and start the server.
So I split it up and used the `console.sh` command to run the server.
Not ideal, but good enough for now to test it.