Simplify JavaScript webpack app
Auto generate the release name, so you can't forget.

Take out the development server setup because yes it will make app
development easier, but it won't actually submit the sourcemaps for a
new release that way. The release is a static value in the webpack
config on start, you can't change that later.