Use NFS volumes to speed up builds (#18)
Docker volume performance on macOS is not great. Apps that rely on a lot
of file system interaction (reading and writing) are much slower on the
mounted volume than the host system.

Use NFS volumes to speed up the mounted volumes. NFS seems much faster
for apps such as Node.js, as per a very basic speed test comparison:

Docker volume vs NFS volume speed test running `yarn install` in
`appsignal-javascript/packages/javascript`, lower is better:

- standard volume: 307.25s
- NFS volume: 122.22s

No significant increase of build time for `yarn build` for the same
package, lower is better:

- standard volume: 33.59s
- NFS volume: 28.97s

Closes #17

## Installation

On macOS there are some additional steps required to set up NFS volumes
on the system. I've added this in an automated way to the
`app:prepare_nfs` task, which runs before `app:up`. This way a user
shouldn't have to bother with these steps manually or run a separate
task beforehand.

It is based on this gist:
https://gist.github.com/seanhandley/7dad300420e5f8f02e7243b7651c6657
From this article:
https://sean-handley.medium.com/how-to-set-up-docker-for-mac-with-native-nfs-145151458adc