Add typescript angular test app
Used to debug issue
https://github.com/appsignal/appsignal-javascript/issues/465

It is set up in the same way as the javascript-webpack test app
https://github.com/appsignal/test-setups/pull/16

It's a little non standard setup, but I cannot wait for the container to
restart when running the server in `boot.sh`. It takes 10+ minutes on my
machine to build all the integrations, link them and start the server.
So I split it up and used the console.sh command to run the server. Not
ideal, but good enough for now to test it.