Remove processmon from checked in files
We don't need a 5MB binary in the history, especially if it gets
updated every time someone runs the `rake global:install_processmon`
task. Instead, build it automatically when it's not present. This
doesn't require any actions by the user, it's done automatically, but
the user can always run `rake global:install_processmon` manually to
update the binary.