Use mono project in Node.js test setups
Use our own mono repo tool for Node.js apps:
https://github.com/appsignal/mono/

Currently relies on the branch of PR
https://github.com/appsignal/appsignal-nodejs/pull/394

It requires a bit of extra work to update to npm 7, which is required
for mono. This because of how the node Docker image works, you need to
change the location of the global npm packages install location, then
add that location to the $PATH and finally install npm globally to
update it. Without these extra steps `npm install -g npm` would fail on
permission issues.