Remove NFS volumes from apps
While NFS volumes speed up the performance, they cause issues with the
agent build. The setup also uses mounts, but they don't work with this
config.

We started using processmon for apps to auto restart apps based on
changes in the integration or the app, so the performance decrease
should be less noticeable.