Add Sinatra app with Puma server
Test the phased restart system.
This app doesn't use processmon and it's testing the Puma phased restart
system instead. If processmon would restart the app every time on
changes it would become more difficult to verify if the app was really
restarted properly. The problem now is that the AppSignal gem breaks
this phased restart process.
https://docs.appsignal.com/support/known-issues/puma-phased-restart.html