Remove strict version locks from Rails test app
The latest ruby 2.7 Docker image is now 2.7.4 and the Gemfile is locked
on 2.7.2. Remove this lock so that reinstalling Docker, and fetching the
new 2.7 image, doesn't break the app on installation.

We could also make the Dockerfile "FROM" statement more strict, which I
would do for production apps, but these are test apps and it would be
nice to use the latest version once it comes out.