Add basic vanilla JavaScript test app (#30)
Use esbuild to build the JavaScript assets to keep the setup simple.

We still need a better way to set the front-end API key. For now I've
added another credentials file.

It's used to test the upcoming change to the JavaScript library to
support a callback function to modify the span before sending it to
AppSignal APIs.