Fix Javascript test setups
This commit fixes the `react`, `typescript-angular` and `webpack`
test setups, using the `vanilla` test setup as reference.

In order to build the integration from source, `mono` must be
installed. This requires changes to the `Dockerfile`s, which are
now unified across all the Javascript test setups.

Test setups that had their commands as `boot.sh` and `console.sh`
had those updated to `run` and `console` respectively, and the
`run` command has been updated to use `mono` in the way that the
`vanilla` test setup does. The way these test setups work has
otherwise been maintained as-is: specifically, that
`typescript-angular` and `webpack` don't actually start their
servers on `app:up`, but on `app:console`.

On the `typescript-angular` test setup, the timeout for submitting
the sourcemaps had to be increased. It is unclear to me why that
is the case.