Fix Next.js test setup
The Next.js test setup specified `~> 10.0` for its dependency on
the `next` package, which limits it to versions of the form
`10.0.x`. However, our Next.js integration package currently
requires `>= 10.2.3`.

This commit changes the test setup's specified dependency to
`^ 10.0`, so that it accepts any versions of the form `10.x.x`.