Add test suite for server test setups
In order to test that the server-based test setups can boot up
correctly, a test suite is created at `support/server-tests`,
with a single test that ensures that the server responds to a
`GET /` HTTP request with a success status code.

Note that, at this point in time, we are not checking that the
integration with Appsignal within these test setups works.

This test suite is ran as an additional service in each of the
test setups' `docker-compose.yml` files. This service will only
run when the `tests` profile is enabled. The service waits for
the server to start responding to requests before the test
suite starts.

In order to pass this test, test setups were modified as follows:

- `elixir/phoenix-example`: A call to a `slow()` method within
  the index controller, which caused requests to time out, was
  removed.
- `nodejs/express-apollo`: The test setup was modified to use
  `apollo-server-express`. While the test setup was already
  configuring both an apollo server and an express server, only
  the apollo server would be started and listen to requests. An
  index route was added to the express server.
- `ruby/rails-postgres`: A `welcome#index` controller was created
  and configured as the root controller. The configuration was
  modified to allow connections from the `app` host.
- `ruby/rails-sidekiq`: The configuration was modified to allow
  connections from the `app` host.

The `javascript/typescript-angular` and `javascript/webpack` test
setups are not covered by these tests, because, despite being
server-based, they do not start their server on `app:up`, but on
`app:console`.

A `.semaphore/semaphore.yml` file is added describing how these
tests are to be run in CI. The `ruby/linux-arm` test setup is
excluded from running in CI, as the machines running the tests
in CI are x86_64 machines running the Docker daemon directly,
without the virtualization support provided by Docker Desktop.

The server tests have been added to the `docker-compose.yml.erb`
template for creating new test setups.