Add slow and error tests for backend setups
This commit adds two new tests that run against the `/slow` and
`/error` endpoints on the test setups. For the former, it checks
that the request takes at least two seconds to execute, and that
it eventually replies with a success status code. For the latter,
it checks that it replies with an error status code.

These tests only make sense for backend setups, so the tests check
whether they're running against a backend setup before attempting
to run these tests.

A helper is also added, implementing a matcher for response status
codes.