Use Dart Sass instead of Node Sass
Bumping the Elixir container version to 1.13 gets us a newer
version of Alpine, which installs a newer version of Node.js.

The `node-sass` dependency uses node-gyp to compile libsass, and
the version of it in the package.json is not compatible with the
new Node.js version. Updating `node-sass` to a newer version would
force us to update `sass-loader` to a newer version, which would
force us to update `webpack` to a newer version. Updating `webpack`
to a newer version, in turn, implies lots of changes for the
ecosystem of plugins that are bundled into the app.

So, instead of that, I'm swapping out `node-sass` for `sass`, also
known as "Dart Sass", which is apparently the current version of
Sass under development. It does not need to install any native
gizmos, so it's compatible with the new version of Node.js, while
also being compatible with the `sass-loader` version already
present.