Add Rails v6 with Mongo setup
This commits adds a new test setup consisting on Rails v6 + Mongoid.
Didn't use Rails v7 as it isn't supported by Mongoid yet.

The setup includes the test endpoints `slow` and `errors` and a couple
of endpoints that allow you to list, show and create blog posts for
testing purposes.