Pin broken builds to npm version 7
Using npm version 7 seems to avoid a bug where the build for the
`@appsignal/nodejs` package is not ran before `npm install`,
causing the installation to fail as it depends on the Transmitter,
which needs to be transpiled first.

The bug seems similar to [npm/cli#4552][bug], although that seems
to be fixed in npm 8.10.0, and upgrading to that version did not
fix the problem. I have been unable to reproduce the bug outside
of Semaphore CI, which further complicates investigating it.