Configure app port to always be 4001 (#81)
The difference between the Docker compose internal network port and the
port exposed to the outside has tripped me up more than once. For
instance when the container output says something like: "The app is
ready at 127.0.0.0:3000", but it's not.

Let's use the same port everywhere. Configure the PORT env var by
default, because we need to configure the port for some apps manually
anyway, and others listen the this PORT env var.