Disable Git safe directory checks
Since the Alpine Dockerfile does not come with Git, we install it
as part of the Docker build, since our scripts and `mono` both
depend on it.

Due to this, we're installing a newer version of Git, which defaults
to raising a fatal error if the user who owns the repository files
does not exist in this machine.

Since `/integrations` is a volume mounted from the host, this will
always be true for the guest, so this commit disables safe directory
checks.