Add Next.js `appDir` app
This uses the new "app directory" Next.js setup, which internally
changes pretty much everything about how Next.js works.

An app is created with the `npx create-next-app` installer, then
the `instrumentation.js` and `appsignal.cjs` are added to make the
instrumentation happen, as well as enabling instrumentation in
`next.config.js`.

Other routes have been created to pass the tests or to enable
testing different aspects of the integration.