Specify specific Ruby Dockerfile base images
I think our build is broken because it previously used the bullseye
release of Debian, and since bookworm has come out, the Ruby images were
updated to use that release by default. By specifying exactly which
version we want, fix the immediate problem with the build.

This doesn't resolve #117, I think we should still upgrade the Rails
versions to 6.1, but at least the build works again with this change.