Add Rails 6 + Shakapacker example app
During support made an app for a Rails app using our frontend
integration and Shakapacker for asset management and sourcemap
generation.

The `bin/rails sourcemaps` task generates the assets and uploads the
sourcemaps to AppSignal.

It does require a separate APPSIGNAL_FRONTEND_API_KEY set in
`appsignal_key.env` in the app's directory.