Create log file before tailing it
If the log file doesn't exist yet, I don't want to have to keep retrying
the command until it does. This extra `touch` makes sure the file exists
and we can tail it.

I tried to combine the commands into one `docker-compose exec`, but then
you need to specify the shell the container uses, which we don't know
from the outside.
I tried to rely on the `APPSIGNAL_LOG_PATH=/tmp` env var set in
`appsignal.env`, but had the same problem.