Delete trailing `/` from app name
Since the test setup app names map to the folder structure, we can
make use of the shell's autocomplete to find the right app, which
is neat.

But the shell will add a `/` to the end when it's a full match,
which makes Docker sad when running `app:down`. This fixes that.