Use latest Prisma instrumentation version
For some reason, the `package.lock` in the test setup has its own
idea of which version of `@prisma/instrumentation` should be used,
which overrides the one in the integration's own `package.lock`.

By running `npm link` with `--save --package-lock`, we ensure that
the test setup uses the versions of the integration's dependencies
that are defined in the integration's package lock.

By running it _before_ `npm install`, we make sure that the test
setup does not install its own versions of the dependencies.

I am not sure if this makes sense, in general, or for other test
setups as well. I lost a few sanity points in figuring this out.