Add new Next.js app
This uses the App Router, but that's the default now, so no config
needed like in `nextjs-appdir`.

This app more closely resembles a new Next.js app. The one a user in
support is having issues with.

Add a system to choose to run the Next.js production mode instead of the
development mode without needing a whole new app, using Docker build
args.