Log to both STDOUT and AppSignal in Sidekiq app
Make development/testing easier by logging to both the terminal/STDOUT
and AppSignal (what we also want to test if it works). Without this
change it would only log to AppSignal, which is delayed and not part of
a local dev workflow.

Use ActiveSupport::Logger.broadcast to wrap loggers and send logs to
multiple backends.

Move the Rails logger config from `config/environment.rb` to
an initializer as documented in
https://github.com/appsignal/appsignal-docs/pull/1453.