Remove (unintentional?) WSGI instrumentation
We're manually applying the WSGI instrumentation in `wsgi.py` in
the Django test setup, alongside using `appsignal.start()` on
`manage.py`. This results in event timelines that contain a WSGI
span around the Django span.

For now, remove this, as this is not what we tell users to
configure in our documentation.