Add app demo command
Make it easy to run the demo command for integrations without having to
run a bash/console command, navigate to the app directory and remember
what the demo command is for an app.

I don't expect every app to have a demo command, but this makes it
possible to easily add one to apps.