Add Ecto models and preload endpoint
Implement a series of Ecto models with associations and create an
endpoint that performs a preload query against them.

This follows the [Elixir School Ecto lessons][tutorial] step by step
all the way to and including the "Associations" section. The sample
queries at the end of the lesson that show how to create data
associatively have been turned into an additional migration. The
"distribution" model has been omitted.

The `/preload` endpoint reproduces the bug where parallel queries
in preloads are not sent to AppSignal.

[tutorial]: https://elixirschool.com/en/lessons/ecto/associations