Use Rack's EventHandler for transaction management
The EventHandler's `on_finish` event seems to be the most reliable way
to close the AppSignal transaction regardless if there was an error or
not.

This also allows us to track the entire request, not just wherever we
were on the middleware stack, assuming our EventHandler is at the top of
the middleware stack.

This change also requires the AppSignal Ruby gem's BodyWrapper class to
no longer call `Appsignal::Transaction.complete_current!` anywhere.