Add Express + BullMQ test setup
Add a test setup that uses Express + BullMQ to enqueue and execute
jobs. Supports enqueueing the job to be processed by a worker in the
same process, as well as by a worker in a separate process.

This uses the before build hook to clone the BullMQ OpenTelemetry
instrumentation repository. This is a temporary step -- on the actual
release, the BullMQ instrumentation will be a dependency of the
AppSignal for Node.js integration.

Similarly, the instrumentation is currently added using the
`additionalInstrumentations` config option -- this will be removed
for the actual release, as the instrumentation will be automatically
added by the AppSignal for Node.js integration.

Currently, my fork of the repository is the one being cloned, but the
same setup could be used if the fork is moved to the AppSignal GitHub
account.