Do not lock AppSignal version for Next.js
The Next.js test setups, due to a Webpack limitation, do not use the
source code of the integration at `nodejs/integration`, instead
installing it from the NPM servers.

Due to this, it makes sense for its bound to be set to `latest`, to
prevent accidentally testing against a very old version of the
integration.