Add metadata to collector test apps (#267)
Add AppSignal magic attributes for:

- Request parameters
- Request session data
- Function parameters

And configure the resource with filtering for these attributes.

The values need to be stored as JSON because OpenTelemetry doesn't
support setting a keylist value as an attribute value.

For Elixir I set the config options as CSV values, because I couldn't
get a list to work. We support both in the collector, but I prefer the
list type values.

[skip review]