Fix context passing between Gin and OTelSQL
OTelSQL expects to be given a [`context.Context`][context], which is
part of the Go standard library, and is essentially an untyped
key-value map that cosplays as a task cancellation mechanism on
weekends.

OpenTelemetry uses the `context.Context` (in its untyped key-value map
aspect) [as its trace context propagation mechanism][otel], similar
to how it uses its own globally-retrievable trace context types in
other languages. This is a sensible design choice, as most Golang
web frameworks expect to pass a context around.

Now, the `go-gin` web framework has a [`gin.Context`][gin], which _is_
a `context.Context`, in that it implements the same interface, but
is also... a bunch of other stuff. Specifically, it stores information
about the request and the response -- and the request object it stores
has _its own separate `context.Context`_ within it.

The `otelgin` library [writes the OpenTelemetry context][otelgin]
information for the request, not to the `gin.Context` itself, but to
the request context within it -- to the `context.Context` at
`ctx.Request.Context()`, where `ctx` is the `gin.Context`.

Therefore, we need to pass `ctx.Request.Context()` to the database
handle that has been initialised by OTelSQL`, as that is the context
that contains the OpenTelemetry information required to set the
parent for the SQL spans.

[context]: https://pkg.go.dev/context#Context
[otel]: https://github.com/open-telemetry/opentelemetry-go/blob/main/trace/context.go
[gin]: https://github.com/gin-gonic/gin/blob/v1.10.0/context.go
[otelgin]: https://github.com/open-telemetry/opentelemetry-go-contrib/blob/instrumentation/github.com/gin-gonic/gin/otelgin/v0.59.0/instrumentation/github.com/gin-gonic/gin/otelgin/gintrace.go#L86-L87