Add Java + Spring Boot test setup
This is at its core a re-arranged version of the `spring-native` app
from the open-telemetry/opentelemetry-java-examples repository:

https://github.com/open-telemetry/opentelemetry-java-examples/tree/main/spring-native

In the original repository, this app is a sub-project of a greater
Gradle project. I have kept this structure, which accounts for some
of the weirdness in the folder structure, but removed all other
projects from the original repo.

I have also modified so that, rather than using Gradle from the host
to build a Docker image that is then ran from Docker Compose, the
entire build and run happens from within Docker, removing the need
for a Java toolchain.

The OpenTelemetry configuration has been modified to set the required
resource attributes for the AppSignal collector to work, and the
OpenTelemetry collector has been replaced with the AppSignal
collector.

An environment variable to disable the log exporter has been added,
removing noise about the missing `/v1/logs` endpoint in the AppSignal
exporter.