Add React + Vite collector test setup
Used Vite to create a React 19 app, alongside the OpenTelemetry
web plugins for instrumenting document loads and React components.

It creates "traces" that represent different stages of the component
lifecycle. So, for example, a trace is created when the component is
mounted, going over different bits of the methods that are used to
handle this on React, then another trace is created when it is updated,
and a third trace is created when it is unmounted. These traces are
not connected to each other, via links or otherwise.

Similarly, when a component mounts another component, these do not
constitute a single trace, but two disconnected ones. The same goes
for the document load trace, which is disconnected from the React
component traces.

The integration predates, and does not support, the not-so-new style
of function-based components. In order to instrument these components,
it was necessary to convert them to old-style, class-based components.

The integration also does not officially support React 17 or greater.
The `--force` flag is passed to `npm install` in order to force it to
work with React 19.

I added an instrumented component that throws an error (along with an
error boundary that stops it from crashing the page) but no error is
reported automatically. One could, of course, use the error boundary
component to manually report an exception event, like our front-end
integration does.

Added a small NGINX-based CORS proxy to work around CORS issues. If
we're interested in supporting this, we may want to have the collector
emit the right CORS headers.

Metrics do not seem to be supported.