Extract Docker setup from vendor folder
Bafflingly, Sail ships the Dockerfile for the Laravel project as
part of its own package. This means that you need to first run
`composer install` locally for the Dockerfile to actually exist.

Copy the Dockerfile and the files it requires to a separate folder
that is committed alongside the app, and install the dependencies
when the app starts.