Install OpenTelemetry dependencies
Install the dependencies required by the OpenTelemetry extension, as
well as the Laravel autoinstrumentation.

Downgrade Laravel to version 11. This is a temporary workaround as
there exists [a PR][pr] in the instrumentation repository that will add
support for Laravel 12.

Set the required environment variables for OpenTelemetry to work with
the AppSignal collector.

Fix an issue where migrating the database would fail if it took place
while the database was being restarted.

[pr] :https://github.com/opentelemetry-php/contrib-auto-laravel/pull/7/files