Fix Elixir apps failing to install dependencies
We're getting this error on several Elixir apps.

Bump the hackney and certifi packages to fix the issue.

```
Resolving Hex dependencies... 02:00
Resolution completed in 0.296s 02:01
Because "the lock" depends on "hackney 1.20.1" which depends on "certifi ~> 2.12.0", "the lock" requires "certifi ~> 2.12.0". 02:01
And because "your app" depends on "the lock", "certifi ~> 2.12.0" is required. 02:01
So, because "your app" depends on "certifi ~> 2.14", version solving failed. 02:01
** (Mix) Hex dependency resolution failed
```