Configure laravel with code from .env
In bootstrap/app.php, we manually call the LoadEnvironmentVariables
bootstrapper on the app instance, then initialise the OpenTelemetry
instrumentation, then return the app.

This makes it possible to configure AppSignal with the same environment
variables that the Laravel application will use, while also
initialising OpenTelemetry early enough that Laravel routes are
instrumented.

The instrumentation file itself is generic and should work with any
other PHP framework, though this of course hasn't been tested. It may
be worth building a test setup with Symfony, or if not, one using Slim,
and optionally Dotenv.