Fix `minimatch` version and add lockfile
Use the `resolutions` field to pick a compatible version of the
`minimatch` library, preventing the `@types/minimatch` stub package
from triggering the installation of a version that is not compatible
with the test setup's Node version, which crashes the installation.

Commit the `yarn.lock` lock files for the test setups.