Configure custom error handler route for Spring
In spring, error get automatically forwarded to the `/error` route. This
is a route we test against and would only work if you access it through
`curl` directly for some reason.

Remap the route to `/error-handler`, a route that doesn't exist on
purpose, so that calling another route like `/error_with_cause` doesn't
also trigger the error in the `/error` route.