Fix `winston` version mismatch
The integration uses a more recent version of `winston`, which requires
a newer version of `logform`, which in turn has added a symbol index
signature to the `TransformableInfo` interface.

This change updates the versions of `winston` to fix the mismatch, and
`@types/node` to stop `tsc` from complaining about missing `fetch`.